+++ CHANGELOG.md
@@ -1,3 +1,20 @@
+## CAPTURE 1.0.0 (January 16, 2026) ##
+* Change license from MIT to GPLv3
+* Add the `cap verify` command
+* Show environment name for `cap run`
+* Check that environment files exist for cap run -e
+* Fix main branch remote tracking
+* Require the env, run, and verify commands to be ran from the project root directory.
+* Rename bin/conda to bin/env
+* Add --source-file-name option to cap_data_download
+* Add -e,--environment options to `cap env` command
+* `cap env` shows all variables created in the CAPTURE environment setup
+* Add job id to log file names of non-array jobs
+
+:warning: **WARNING** This is a breaking change!
+* Add slurm option to `cap run`. The default has changed from running the script
+  with Slurm sbatch to running it in the current terminal session.
+
 ## CAPTURE 0.8.6 (October 21, 2025) ##
 * Fix cap_container failure on Slurm array jobs